fix(build-audit): move utils into separate dir
