+++ libs/feature/audit-builder/src/lib/audit-builder/audit-globals.component.ts
@@ -8,7 +8,7 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { RxIf } from '@rx-angular/template/if';
 import { RxFor } from '@rx-angular/template/for';
 import { DEVICE_TYPE } from '../schema/audit-builder.constants';
-import { ToTitleCasePipe } from './toTitleCase.pipe';
+import { ToTitleCasePipe } from '../utils/toTitleCase.pipe';
 import { AuditBuilderService } from './audit-builder.service';
 
 @Component({
